+++ hc/accounts/tests/test_project.py
@@ -3,7 +3,7 @@ from django.core import mail
 from django.conf import settings
 from django.test.utils import override_settings
 from hc.test import BaseTestCase
-from hc.accounts.models import Member
+from hc.accounts.models import Member, Project
 from hc.api.models import TokenBucket
 
 
@@ -87,6 +87,28 @@ class ProjectTestCase(BaseTestCase):
             " Alice's Project on %s" % settings.SITE_NAME
         )
         self.assertHTMLEqual(mail.outbox[0].subject, subj)
+
+    def test_it_adds_member_from_another_team(self):
+        # With team limit at zero, we should not be able to invite any new users
+        self.profile.team_limit = 0
+        self.profile.save()
+
+        # But Charlie will have an existing membership in another Alice's project
+        # so Alice *should* be able to invite Charlie:
+        p2 = Project.objects.create(owner=self.alice)
+        Member.objects.create(user=self.charlie, project=p2)
+
+        self.client.login(username="alice@example.org", password="password")
+        form = {"invite_team_member": "1", "email": "charlie@example.org"}
+        r = self.client.post(self.url, form)
+        self.assertEqual(r.status_code, 200)
+
+        q = Member.objects.filter(project=self.project, user=self.charlie)
+        self.assertEqual(q.count(), 1)
+
+        # And this should not have affected the rate limit:
+        q = TokenBucket.objects.filter(value="invite-%d" % self.alice.id)
+        self.assertFalse(q.exists())
 
     @override_settings(SECRET_KEY="test-secret")
     def test_it_rate_limits_invites(self):